Improve CLI workflow with npm scripts
At this point the structure of this starter package is finally taking
shape. It looks like the best solution for now is to use Gulp as a
mini-cli (with tasks invoked through NPM scripts). This should impose
the least amount of cognitive overhead on the user. All they have to do
is:

1. Make sure they have hugo and git installed on their system
2. Download the git repo for the starter package
2a (optional). Run `git remote rm origin` and replace with their own
3. Run `npm install` inside the project folder
4. Run `npm run setup`, `npm run dev`, and `npm run build` as needed.