Fixes bibliography for cleaner matching
This fixes the issue that bibliography entries were being matched too
greedily. For example on page 4, entries for 4, 14, 40, 41 etc were all
being called. This also allows a single bibliographic entry to be
assigned to multiple pages in a publication.

So that users can connect references in `bibliography.yml` to pages
with a simple syntax of `reference_for: “1, 2, 3”` (like they do for
the  `q-figure-group` shortcode) rather than more complicated yam array
syntax Hugo seems to need, the template now takes the user-provided
strings, converts them to arrays, and loops through them to compare to
the page ‘id`.

The template currently runs the same function beforehand to determine
if there are any relevant entries in `bibliography.yml`. If there  are
not, no “Bibliography” navigation link or header are added to the page.
There is probably a cleaner way of doing this rather than repeating the
same template function, but that’s for another time