Add speculative no-js progressive enhancement
Elements that have some dependency on JavaScript to look or operate
correctly are tagged with a `no-js` class in the markup. If JavaScript
is active, the class is removed and the site functions as expected. If
JS is not active, the class allows us to restyle elements accordingly,
with the goal that the no-JavaScript mode feels as clean, complete and
intentional as possible.