Fix background image so it shows up on GH Pages
This required some re-organization. The cover background image should be
thought of as an "asset", not as part of the book's content. So the css
that controls the background image of the cover page should look for an
image with the path of "../img/cover_bg.jpg". This will always reference
the top-level img folder no matter where the site is built (locally or
on GH Pages).

Because Webpack is controlling the entire asset pipeline now (including
CSS), it will blow up and throw an error if an image matching the name
of cover_bg.jpg does not exist at compile time. So a placeholder image
has been added to the source/img folder of the theme. This could be
overridden at the content level with an image in static/assets/img (I've
provided one as an example in this commit, but it is just the same image
again).