WIP: Photoswipe integration
This gets basic photoswipe integration for figures mostly working.
However, the way "figures" work in this publication is more complex than
in previous projects. Figures can contain other figures, they can also
contain tables, etc. We need to come up with a way to designate
"bottom-level" figures which we want to target for displaying inside a
lightbox and name them accordingly.

For example, a solo figure with an image and caption should get this
designation. But a group figure with multiple images should not - each
sub-figure with a single image child element should be loaded instead.
HOWEVER the shared caption from the parent figure group should probably
be used for both.

We will need to figure out a way to deal with questions like this before
this feature can be merged in.