Handle various forms of figures/fig captions in Photoswipe (#14)
This commit updates the lightBox function and the global event handlers to
better accommodate the figure & caption structure in this project for Photoswipe
integration.

Some helper functions have been introduced to provide the necessary logic to do
this since there are several different ways that figures and figure captions
might be arranged in this publication. These functions use a recursive
process (contained in the searchUpTree() function); when Photoswipe is
triggered, it will go through each slide and test first the figure element, and
then any of its parent elements for a <figcaption> tag to use for the slide's
caption text.