Add basic table handling to lightbox (#15)
Photoswipe has the ability to handle HTML in addition to regular images. The
library assumes that any HTML will be very simple so you are on your own for
implementing anything more complicated.

This commit adds a basic ability to go full-screen when table wrapper divs are
clicked, and they show up in the standard order of slides. Captions, etc. work
as normal.

Styling is pretty simple right now - the table sits at the center of the screen
on a white background. For smaller tables this works fine; for the really large
and complex ones, we're going to need some kind of scrolling behavior (so PSWP
will need to have its own scrolling listeners disabled so they do not
interfere).