Update search index generator template
The Go template that produces the search index has been overhauled in order to
handle the special case of "absract-only" pages. These are still classified as
"Essay" type but they have no content, only `abstract` frontmatter data.

The desired behavior for search is that on these pages, the user should be taken
to the section-head page that this abstract appears in if they choose a result
of this type.

In Go, the different cases (normal vs abstract-only pages) are handled via
conditional assignment of block-scoped variables in the local .Scratch (as
opposed to the page-level $.Scratch).

The additional logic comes into play if the length of a given page's .Content is
0 (i.e. empty). In that case, the abstract content is used instead, and instead
of using the normal .RelPermalink for the page, the URL data is set as the
relref of the page's parent section (figured out with a crude hack for now -
taking the name of the $page.Section value and adding "/index.md" on the end).