Basic Search page now working
This needs a pass for QA as well as UI, but the basic functionality is now
working. Previous projects set this feature as something available on all pages;
in this book the user needs to navigate to the /search page to access this
screen. This simplifies things both in terms of CSS layout, but also performance
issues (the search index is 1MB, and now it doesn't need to be made available on
all pages, just the search-specific page).

A "search.md" page has been added to the contents folder. Any content here
will not show up, but we need to tell Hugo to build a page using the search
layout at this URL.

The global `baseof.html` template has been modified so that the search link now
appears in all of them (not the search button and results partial from before).
A new search/single.html layout has been added to the base theme which handles
this content instead.

Finally, the main UI.js file has been re-worked a bit to accommodate the new
behavior.