Use a CSS solution to deal with anchor scroll
The anchorScroll() function was causing some errors, and having a JS function
step in and intercept every link click event is a heavy solution that seems best
avoided.

This commit disables the anchorScroll() function and replaces it with a
CSS-based solution for adjusting scroll height to account for the fixed header.